# public/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ShadowSpark Autonomous Revenue API",
    "version": "2.0.0",
    "description": "API documentation for the ShadowSpark autonomous infrastructure."
  },
  "servers": [
    {
      "url": "https://shadowspark-tech.org/api",
      "description": "Production Server"
    }
  ],
  "paths": {
    "/leads/sync": {
      "post": {
        "summary": "Synchronize lead from external source",
        "security": [{ "ApiKey": [] }],
        "responses": {
          "200": { "description": "Lead synchronized" }
        }
      }
    },
...
    "/webhooks/paystack": {
      "post": {
        "summary": "Process Paystack webhooks",
        "description": "Verifies the HMAC-SHA512 signature, ignores duplicate references, updates the lead tier, and records a payment_received SystemEvent.",
        "responses": {
          "200": { "description": "Webhook accepted" },
          "400": { "description": "Missing reference or leadId" },
          "401": { "description": "Invalid signature" },
          "500": { "description": "Webhook processing failed" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-sync-secret"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "authjs.session-token"
      }
    }
  }
}